backend/requirements.txt
fastapi
uvicorn
pillow
requests
python-multipart
python-dotenv
supabase
PyJWT
openai
